Vinetime - frontend rebuild after Figma issue
I built the VineTime design in Figma myself and tried to connect it with React at first. Because the project setup was with Thymeleaf, I used Claude to help me convert and restyle everything in plain HTML, CSS, and JS. Learnt a lot about frameworkds and how easily things can clash in shared projects.